requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# or
# $ pip install -r <this file>
# platform: linux-64
_libgcc_mutex=0.1
_openmp_mutex=5.1
blas=1.0
bottleneck=1.3.7
brotli=1.0.9
brotli-bin=1.0.9
bzip2=1.0.8
c-ares=1.19.1
ca-certificates=2024.3.11
cairo=1.16.0
contourpy=1.2.0
cycler=0.11.0
cyrus-sasl=2.1.28
dbus=1.13.18
eigen=3.4.0
expat=2.5.0
ffmpeg=4.2.2
fontconfig=2.14.1
fonttools=4.25.0
freetype=2.12.1
giflib=5.2.1
glib=2.78.4
glib-tools=2.78.4
gmp=6.2.1
gnutls=3.6.15
graphite2=1.3.14
gst-plugins-base=1.14.1
gstreamer=1.14.1
h5py=3.9.0
harfbuzz=4.3.0
hdf5=1.12.1
icu=58.2
jpeg=9e
kiwisolver=1.4.4
krb5=1.20.1
lame=3.100
lcms2=2.12
ld_impl_linux-64=2.38
lerc=3.0
libbrotlicommon=1.0.9
libbrotlidec=1.0.9
libbrotlienc=1.0.9
libclang=14.0.6
libclang13=14.0.6
libcups=2.4.2
libcurl=8.5.0
libdeflate=1.17
libedit=3.1.20230828
libev=4.33
libevent=2.1.12
libffi=3.4.4
libgcc-ng=11.2.0
libgfortran-ng=11.2.0
libgfortran5=11.2.0
libglib=2.78.4
libgomp=11.2.0
libiconv=1.16
libidn2=2.3.4
libllvm14=14.0.6
libnghttp2=1.57.0
libopenblas=0.3.21
libopus=1.3.1
libpng=1.6.39
libpq=12.17
libprotobuf=3.20.3
libssh2=1.10.0
libstdcxx-ng=11.2.0
libtasn1=4.19.0
libtiff=4.5.1
libunistring=0.9.10
libuuid=1.41.5
libvpx=1.7.0
libwebp=1.3.2
libwebp-base=1.3.2
libxcb=1.15
libxkbcommon=1.0.1
libxml2=2.10.4
libxslt=1.1.37
lz4-c=1.9.4
matplotlib-base=3.8.0
munkres=1.1.4
mysql=5.7.24
ncurses=6.4
nettle=3.7.3
nspr=4.35
nss=3.89.1
numexpr=2.8.7
numpy=1.26.4
numpy-base=1.26.4
opencv=4.6.0
openh264=2.1.1
openjpeg=2.4.0
openssl=3.0.13
packaging=23.2
pandas=2.2.1
pcre2=10.42
pillow=10.2.0
pip=23.3.1
pixman=0.40.0
pyparsing=3.0.9
python=3.10.14
python-dateutil=2.8.2
python-tzdata=2023.3
pytz=2023.3.post1
qt-main=5.15.2
qt-webengine=5.15.9
qtwebkit=5.212
readline=8.2
scipy=1.12.0
seaborn=0.12.2
setuptools=68.2.2
six=1.16.0
sqlite=3.41.2
tk=8.6.12
tzdata=2024a
wheel=0.41.2
x264=1!157.20191217
xz=5.4.6
zlib=1.2.13
zstd=1.5.5
